fixes based on previous PR feedback
